requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.22.0
pydantic==2.7.1
python-multipart==0.0.9
python-socketio
grpcio

passlib==1.7.4
passlib[bcrypt]
PyJWT[crypto]

requests==2.32.2
aiohttp==3.9.5
httpx

# AI libraries
openai
anthropic
google-generativeai
vertexai

# Database
pymongo
peewee
SQLAlchemy
boto3
redis
sqlmodel
chromadb
psycopg2-binary

# Observability
langfuse
ddtrace
opik

# ML libraries
torch
numpy
pandas

xgboost
scikit-learn

# NLP libraries
sentence-transformers
transformers
tokenizers
nltk
tiktoken

# Image processing
Pillow
opencv-python

# Visualization
matplotlib
seaborn

# Web scraping
selenium
playwright
beautifulsoup4

# Haystack
haystack-ai
chroma-haystack

# Langchain
langchain
langchain-text-splitters

# Neo4j
neo4j
neo4j-graphrag
neo4j-graphrag[sentence-transformers]

# Llama Index for RAG
llama-index
llama-index-llms-ollama
